+++ zorka-core/src/main/java/com/jitlogic/zorka/core/UtilLib.java
@@ -19,14 +19,19 @@ package com.jitlogic.zorka.core;
 import com.jitlogic.zorka.common.util.JSONWriter;
 import com.jitlogic.zorka.common.util.ObjectInspector;
 import com.jitlogic.zorka.common.util.StringMatcher;
+import com.jitlogic.zorka.common.util.TapInputStream;
+import com.jitlogic.zorka.common.util.TapOutputStream;
 import com.jitlogic.zorka.common.util.ZorkaUtil;
+import com.jitlogic.zorka.core.util.Base64;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.regex.Pattern;
 
 public class UtilLib {
 
@@ -130,4 +135,50 @@ public class UtilLib {
         return new JSONWriter().write(obj);
     }
 
+    public Pattern rePattern(String pattern) {
+        return Pattern.compile(pattern);
+    }
+
+    public boolean reMatch(Pattern pattern, String s) {
+        return pattern.matcher(s).matches();
+    }
+
+    private static final int BUF_SZ = 4096;
+
+    public int ioCopy(InputStream is, OutputStream os) throws IOException {
+        int bufsz = Math.min(is.available(), BUF_SZ);
+        final byte[] buf = new byte[bufsz];
+        int n = 0, total = 0;
+        n = is.read(buf);
+        while (n > 0) {
+            os.write(buf, 0, n);
+            total += n;
+            n = is.read(buf);
+        }
+        return total;
+    }
+
+    public long min(long a, long b) {
+        return Math.min(a, b);
+    }
+
+    public long max(long a, long b) {
+        return Math.max(a, b);
+    }
+
+    public byte[] clipArray(byte[] src, int len) {
+        return ZorkaUtil.clipArray(src, len);
+    }
+
+    public String base64(byte[] buf) {
+        return Base64.encode(buf, false);
+    }
+
+    public TapInputStream tapInputStream(InputStream is, long init, long limit) {
+        return new TapInputStream(is, (int)init, (int)limit);
+    }
+
+    public TapOutputStream tapOutputStream(OutputStream os, long init, long limit) {
+        return new TapOutputStream(os, (int)init, (int)limit);
+    }
 }